backend/requirements.txt
fastapi
uvicorn
pillow
python-jose[cryptography]
passlib[bcrypt]
sqlite3
